vmtotal broke on a machine with 32M of mem
SCCS-vsn: sys/sys/vmmeter.h 6.4

+++ usr/src/sys/sys/vmmeter.h
@@ -3,7 +3,7 @@
  * All rights reserved.  The Berkeley software License Agreement
  * specifies the terms and conditions for redistribution.
  *
- *	@(#)vmmeter.h	6.3 (Berkeley) %G%
+ *	@(#)vmmeter.h	6.4 (Berkeley) %G%
  */
 
 /*
@@ -57,15 +57,15 @@ struct vmtotal
 	short	t_pw;		/* jobs in page wait */
 	short	t_sl;		/* jobs sleeping in core */
 	short	t_sw;		/* swapped out runnable/short block jobs */
-	int	t_vm;		/* total virtual memory */
-	int	t_avm;		/* active virtual memory */
-	short	t_rm;		/* total real memory in use */
-	short	t_arm;		/* active real memory */
-	int	t_vmtxt;	/* virtual memory used by text */
-	int	t_avmtxt;	/* active virtual memory used by text */
-	short	t_rmtxt;	/* real memory used by text */
-	short	t_armtxt;	/* active real memory used by text */
-	short	t_free;		/* free memory pages */
+	long	t_vm;		/* total virtual memory */
+	long	t_avm;		/* active virtual memory */
+	long	t_rm;		/* total real memory in use */
+	long	t_arm;		/* active real memory */
+	long	t_vmtxt;	/* virtual memory used by text */
+	long	t_avmtxt;	/* active virtual memory used by text */
+	long	t_rmtxt;	/* real memory used by text */
+	long	t_armtxt;	/* active real memory used by text */
+	long	t_free;		/* free memory pages */
 };
 #ifdef KERNEL
 struct	vmtotal total;
